requirements.txt
Pillow>=3.4.2
bitarray=0.8.1
click>=6.7
pydub>=0.18.0